+++ opentelemetry-sdk/src/metrics/internal/aggregate.rs
@@ -1,4 +1,4 @@
-use std::{marker, sync::Arc};
+use std::{marker, sync::atomic::AtomicU64, sync::Arc};
 
 use once_cell::sync::Lazy;
 use opentelemetry::KeyValue;
@@ -16,7 +16,7 @@ use super::{
     Number,
 };
 
-const STREAM_CARDINALITY_LIMIT: u32 = 2000;
+static STREAM_CARDINALITY_LIMIT: AtomicU64 = AtomicU64::new(2000);
 pub(crate) static STREAM_OVERFLOW_ATTRIBUTE_SET: Lazy<AttributeSet> = Lazy::new(|| {
     let key_values: [KeyValue; 1] = [KeyValue::new("otel.metric.overflow", "true")];
     AttributeSet::from(&key_values[..])
@@ -24,7 +24,12 @@ pub(crate) static STREAM_OVERFLOW_ATTRIBUTE_SET: Lazy<AttributeSet> = Lazy::new(
 
 /// Checks whether aggregator has hit cardinality limit for metric streams
 pub(crate) fn is_under_cardinality_limit(size: usize) -> bool {
-    size < STREAM_CARDINALITY_LIMIT as usize
+    size < STREAM_CARDINALITY_LIMIT.load(std::sync::atomic::Ordering::Relaxed) as usize
+}
+
+/// Set cardinality limit for metric streams
+pub fn set_stream_cardinality_limit(size: u64) {
+    STREAM_CARDINALITY_LIMIT.store(size, std::sync::atomic::Ordering::Relaxed)
 }
 
 /// Receives measurements to be aggregated.